chore: Add Icons component for logo, Twitter, and LinkedIn
A new file, icons.tsx, has been added to the components directory. This file contains the Icons component, which includes functions for rendering the logo, Twitter, and LinkedIn icons. These icons can be used throughout the application for visual representation. This addition improves the modularity and reusability of the codebase.